hexlet_pytest/pyproject.toml
[tool.poetry]
name = "hexlet-pytest"
version = "0.1.0"
description = ""
authors = ["Pavel Soroka <40a.pvl@gmail.com>"]
readme = "README.md"
packages = [{include = "hexlet_pytest"}]

[tool.poetry.dependencies]
python = "^3.10"
schema = "^0.7.5"
faker = "^19.6.2"
pytest = "^7.4.2"
pyyaml = "^6.0.1"
bs4 = "^0.0.1"
pygithub = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
